A-Packages
Organise classes into duck/ and duck/command packages, with src/main/java as source root folder
