--- src/main/java/DuckException.java
@@ -1,5 +1,0 @@
-public class DuckException extends Exception {
-    public DuckException(String errorMessage) {
-        super(errorMessage);
-    }
-}
