--- src/main/java/TaskType.java
@@ -1,6 +1,0 @@
-/**
- * Defines enum such that TaskType can only be Todo, Deadline or Event
- */
-public enum TaskType{
-    ToDos, Deadlines, Events;
-}
