+++ src/main/java/data/duck.txt
@@ -1,0 +1,1 @@
+T | 0 | borrow book
