+++ src/main/java/duck/Duck.java
@@ -14,6 +14,17 @@
 
 // TAKES IN DD-MM-YYYY prints out MMM dd YYYY
 
+package duck;
+
+import duck.command.ByeCommand;
+import duck.command.Command;
+import duck.command.DeadlineCommand;
+import duck.command.DeleteCommand;
+import duck.command.EventCommand;
+import duck.command.ListCommand;
+import duck.command.MarkUnmarkCommand;
+import duck.command.TodoCommand;
+
 public class Duck {
     public static String home = System.getProperty("user.dir");
