+++ src/main/java/duck/DuckException.java
@@ -1,0 +1,16 @@
+package duck;
+
+import duck.command.ByeCommand;
+import duck.command.Command;
+import duck.command.DeadlineCommand;
+import duck.command.DeleteCommand;
+import duck.command.EventCommand;
+import duck.command.ListCommand;
+import duck.command.MarkUnmarkCommand;
+import duck.command.TodoCommand;
+
+public class DuckException extends Exception {
+    public DuckException(String errorMessage) {
+        super(errorMessage);
+    }
+}
