+++ src/main/java/duck/Parser.java
@@ -1,6 +1,18 @@
+package duck;
+
 import jdk.jfr.Event;
 
 import java.util.Scanner;
+
+import duck.command.ByeCommand;
+import duck.command.Command;
+import duck.command.DeadlineCommand;
+import duck.command.DeleteCommand;
+import duck.command.EventCommand;
+import duck.command.ListCommand;
+import duck.command.MarkUnmarkCommand;
+import duck.command.TodoCommand;
+
 
 public class Parser {
     private Ui ui;
