+++ src/main/java/duck/TaskList.java
@@ -1,3 +1,5 @@
+package duck;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.time.LocalDate;
@@ -5,6 +7,15 @@ import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.ArrayList;
+
+import duck.command.ByeCommand;
+import duck.command.Command;
+import duck.command.DeadlineCommand;
+import duck.command.DeleteCommand;
+import duck.command.EventCommand;
+import duck.command.ListCommand;
+import duck.command.MarkUnmarkCommand;
+import duck.command.TodoCommand;
 
 /**
  * Manages a list of all tasks
