+++ src/main/java/duck/command/DeadlineCommand.java
@@ -1,3 +1,14 @@
+package duck.command;
+
+import duck.Duck;
+import duck.DuckException;
+import duck.Item;
+import duck.Parser;
+import duck.Storage;
+import duck.TaskList;
+import duck.TaskType;
+import duck.Ui;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
