+++ src/main/java/duck/command/ListCommand.java
@@ -1,0 +1,19 @@
+package duck.command;
+
+import duck.Duck;
+import duck.DuckException;
+import duck.Item;
+import duck.Parser;
+import duck.Storage;
+import duck.TaskList;
+import duck.TaskType;
+import duck.Ui;
+
+public class ListCommand extends Command{
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage){
+        ui.showList(tasks.toString());
+    }
+
+}
