+++ src/main/java/duck/command/MarkUnmarkCommand.java
@@ -1,3 +1,14 @@
+package duck.command;
+
+import duck.Duck;
+import duck.DuckException;
+import duck.Item;
+import duck.Parser;
+import duck.Storage;
+import duck.TaskList;
+import duck.TaskType;
+import duck.Ui;
+
 public class MarkUnmarkCommand extends Command{
 
     private String fullCommand;
