+++ text-ui-test/runtest.bat
@@ -43,7 +43,7 @@ if exist ACTUAL.TXT (
 
 echo.
 echo [3/4] Compiling source files...
-javac -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\*.java
+javac -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\duck\*.java
 IF ERRORLEVEL 1 (
     echo.
     echo =============== ERROR ===============
@@ -56,7 +56,7 @@ echo [√] Compilation successful
 echo.
 
 echo [4/4] Running tests...
-java -classpath ..\bin Duck < input.txt > ACTUAL.TXT
+java -classpath ..\bin duck.Duck < input.txt > ACTUAL.TXT
 
 FC ACTUAL.TXT EXPECTED.TXT > nul
 if ERRORLEVEL 1 (
